refactor(studio): rename settings route from audit to audit-log
/sites/[siteId]/settings/audit -> /sites/[siteId]/settings/audit-log,
with the sidenav href, page test import, and story pathname updated.
No redirect from the old path: the feature is unreleased and flagged
off, so no links to it exist outside this stack.
