+++ apps/studio/src/features/settings/AuditLogExport/__tests__/auditSettingsPage.test.tsx
@@ -4,7 +4,7 @@ import { ThemeProvider } from "@opengovsg/design-system-react"
 import { render, screen, waitFor } from "@testing-library/react"
 import { beforeEach, describe, expect, it, vi } from "vitest"
 import { UserManagementContext } from "~/features/users"
-import AuditLogExportSettingsPage from "~/pages/sites/[siteId]/settings/audit"
+import AuditLogExportSettingsPage from "~/pages/sites/[siteId]/settings/audit-log"
 import { buildUserManagementPermissions } from "~/server/modules/permissions/permissions.util"
 import { theme } from "~/theme"
 import { RoleType } from "~prisma/generated/generatedEnums"
